Fix bug in chi2 output.
Former-commit-id: eb11c2904719be10ac9380372c87726ff0b00676

+++ reconstruction/dc/src/main/java/org/jlab/rec/dc/banks/RecoBankWriter.java
@@ -305,7 +305,7 @@ public class RecoBankWriter {
 			bank.setShort("Cross2_ID", i,(short) candlist.get(i).get(1).get_Id());
 			bank.setShort("Cross3_ID", i,(short) candlist.get(i).get(2).get_Id());		
 			bank.setShort("status", i, (short) candlist.get(i).status);
-			bank.setFloat("chi2", i, (short) candlist.get(i).get_FitChi2());
+			bank.setFloat("chi2", i, (float) candlist.get(i).get_FitChi2());
 			bank.setShort("ndf", i, (short) candlist.get(i).get_FitNDF());
 		}
 		//bank.show();
@@ -581,7 +581,7 @@ public class RecoBankWriter {
 			bank.setShort("Cross1_ID", i,(short) candlist.get(i).get(0).get_Id());
 			bank.setShort("Cross2_ID", i,(short) candlist.get(i).get(1).get_Id());
 			bank.setShort("Cross3_ID", i,(short) candlist.get(i).get(2).get_Id());
-			bank.setFloat("chi2", i, (short) candlist.get(i).get_FitChi2());
+			bank.setFloat("chi2", i, (float) candlist.get(i).get_FitChi2());
 			bank.setShort("ndf", i, (short) candlist.get(i).get_FitNDF());
 			
 			// save to a separate bank
